Fix reader's association with file extensions in the GUI dialog.
Readers that were available through the GUI file open dialog
were specified based on the client side build, including plugins.
So even though a reader on the server may be available it was
the client that was determining which readers it thought should
be used for opening a file on the server. Thus, if any reader
depended directly on MPI it wouldn't be available on the client
if the client was built in serial. This change fixes that so that
the assocation of file extensions to readers was based on the
server's servermanager XML.

Change-Id: Ie7cf4fd3caaa3bd2dfcf17a33baa5b9c69072af5