Fix saving of unstructrued grids as *.vtk in parallel.
When saving unstructured grids in legacy VTK format in parallel (a way
that's not recommended, and should only be used for small datasets),
ParaView failed since the writer used didn't use the correct type of
filter to merge pieces from different ranks before passing it on to the
legacy writer.

Fixed by replacing the old writer proxy (writers, PDataSetWriter) with
(writers, PDataSetWriterPolyData) and (writers,
PDataSetWriterUnstructuredGrid) which uses different filters for the
post-gather helper based on the type of input dataset.

Addresses #16989.